% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_solver.R
\name{admm_initializer}
\alias{admm_initializer}
\title{ADMM initialization}
\usage{
admm_initializer(
  current_counts,
  weighted_past_counts,
  degree,
  auxi_var = NULL,
  dual_var = NULL
)
}
\arguments{
\item{current_counts}{the current daily infection counts}

\item{weighted_past_counts}{the weighted sum of past infection counts with
corresponding serial interval functions (or its Gamma approximation) as
weights}

\item{degree}{degree of the piecewise polynomial curves to be fitted,
e.g., degree = 0 corresponds to piecewise constant curves}

\item{auxi_var}{auxiliary variable in the ADMM algorithm}

\item{dual_var}{dual variable in the ADMM algorithm}
}
\value{
a list of model initialization with class `admm_initializer`
}
\description{
ADMM initialization
}
