% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_solver.R
\name{admm_solver}
\alias{admm_solver}
\title{ADMM solver}
\usage{
admm_solver(
  current_counts,
  degree,
  dist_gamma = c(1, 1),
  x = double(0),
  lambda = double(0),
  nsol = 100L,
  lambdamin = -1,
  lambdamax = -1,
  lambda_min_ratio = 1e-04,
  maxiter = 10000,
  init = NULL
)
}
\arguments{
\item{current_counts}{the current daily infection counts}

\item{degree}{degree of the piecewise polynomial curve to be fitted,
e.g., degree = 0 corresponds to a piecewise constant curve}

\item{lambda}{a parameter to balance the data fidelity and graphical
smoothness of fitted curves; a greater lambda results in a smoother curve}

\item{maxiter}{maximal number of iteration}

\item{init}{a list of model initialization of class `admm_initializer`}

\item{weighted_past_counts}{the weighted sum of past infection counts with
corresponding serial interval functions (or its Gamma approximation) as
weights}

\item{mu}{a parameter used in the algorithm; use mu = NULL to compute it
using a default method (`get_mu`)}

\item{tol}{tolerance of convergence of primal & dual residuals}
}
\value{
current_counts the current daily infection counts

weighted_past_counts the weighted sum of past infection counts

R_rate: the estimated reproduction rate

convr: if the model converges `convr==TRUE` or not `convr==FALSE`
}
\description{
We should rename this function.
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
admm_solver(
  current_counts = y, weighted_past_counts = rep(1, 10), degree = 1,
  init = admm_initializer(current_counts = y,
  weighted_past_counts = rep(1, 10), degree = 1)
)
}
