% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confband.R
\name{confband}
\alias{confband}
\title{Add confidence bands to estimated Rt curves}
\usage{
confband(object, lambda, level = 0.95, ...)
}
\arguments{
\item{object}{a `poisson_rt` or `cv_poisson_rt` object.}

\item{lambda}{the selected lambda. May be a scalar value, or in the case of
`cv_poisson_rt` objects, `"lambda.min"` or `"lambda.max"`.}

\item{level}{the desired confidence level(s). These will be sorted if
necessary.}

\item{...}{additional arguments for methods. Unused.}
}
\value{
A `data.frame` containing the estimated `Rt` at the chosen `lambda`,
 and confidence limits corresponding to `level`
}
\description{
Create an approximate confidence band for the Rt estimate. Note that the
variance computation is approximate.
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
out <- estimate_rt(y, nsol = 10)
head(confband(out, out$lambda[2]))
head(confband(out, out$lambda[2], level = c(0.95, 0.8, 0.5)))

cv <- cv_estimate_rt(y, nfold = 3, nsol = 30)
head(confband(cv, "lambda.min", c(0.5, 0.9)))
}
