% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv_estimate_rt}
\alias{cv_estimate_rt}
\title{CV for admm}
\usage{
cv_estimate_rt(
  observed_counts,
  degree = 3L,
  dist_gamma = c(2.5, 2.5),
  fold = 3,
  x = NULL,
  lambda = NULL,
  nsol = 100L,
  lambdamin = NULL,
  lambdamax = NULL,
  lambda_min_ratio = 1e-04,
  maxiter = 10000,
  init = NULL
)
}
\arguments{
\item{observed_counts}{}

\item{degree}{}

\item{dist_gamma}{}

\item{fold}{}

\item{x}{}

\item{lambda}{}

\item{nsol}{}

\item{lambdamin}{}

\item{lambdamax}{}

\item{lambda_min_ratio}{}

\item{maxiter}{}

\item{init}{}
}
\description{
CV for admm
}
