% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_calculator.R
\name{delay_calculator}
\alias{delay_calculator}
\title{Delay Calculator}
\usage{
delay_calculator(current_counts, dist_gamma)
}
\arguments{
\item{current_counts}{the current daily infection counts}

\item{dist_gamma}{scale and shape parameter of the discretized Gamma distribution}
}
\value{
list of counts weighted by the discretized Gamma distribution
}
\description{
Delay Calculator
}
\examples{
delay_calculator(c(3,2,5,3,1), c(2,2))
}
