% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.cv_poisson_rt}
\alias{fitted.cv_poisson_rt}
\title{Fitted cv_poisson_rt}
\usage{
\method{fitted}{cv_poisson_rt}(object, which_lambda = NULL, ...)
}
\arguments{
\item{object}{result of cross validation of type `cv_poisson_rt`}

\item{which_lambda}{select which Rt's to output. If not provided, all Rt's
are returned. If it is a list of lambda, in which case
the corresponding Rt estimation will be outputed If it is a string, it must be
either `lambda.min` or `lambda.1se`. If it is `lambda_min`, return
Rt which is generated from the lambda that minimizes the cross validation score.
If it is `lambda.1se`, return Rt which is generated from the lambda whose
corresponding cross validation score is 1 standard error away of the minimal
cross validation score.}

\item{...}{not used.}
}
\value{
Rt's estimated from provided lambda
}
\description{
Fitted cv_poisson_rt
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
cv <- cv_estimate_rt(y, degree = 3, nfold = 2, nsol=30)
f <- fitted(cv)
f <- fitted(cv, which_lambda = cv$lambda[1])
f <- fitted(cv, which_lambda = "lambda.min")
f <- fitted(cv, which_lambda = "lambda.1se")
}
