% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_estimate_rt.R
\name{fold_calculator}
\alias{fold_calculator}
\title{Helper function. Calculate the fold index for each `observed_counts`}
\usage{
fold_calculator(n, nfold)
}
\arguments{
\item{n}{length of the sequence to partition from}

\item{nfold}{Integer. This the number of folds to conduct the leave-kth-out
cross validation. For leave-kth-out cross validation, every kth
observed_counts and their corresponding position (evenly or unevenly
spaced) are placed into the same fold. The first and last observed_counts are
not assigned to any folds. Smallest allowable value is `nfold = 2`. It is
generally not recommended to set `nfold` to be greater than 3.}
}
\value{
a vector of fold index at which the counts are distributed into folds
}
\description{
Helper function. Calculate the fold index for each `observed_counts`
}
