% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildD.R
\name{generate_D}
\alias{generate_D}
\title{Generate a divided difference matrix of order zero}
\usage{
generate_D(n)
}
\arguments{
\item{n}{number of matrix rows}
}
\value{
a (n-1) * n sparse matrix
}
\description{
Generate a divided difference matrix of order zero
}
