% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildD.R
\name{generate_Dk}
\alias{generate_Dk}
\title{Generate a divided difference matrices of order k}
\usage{
generate_Dk(n, k = 1)
}
\arguments{
\item{n}{row dimension}

\item{k}{order of divided difference}
}
\value{
a (n-k-1) * n sparse matrix
}
\description{
Generate a divided difference matrices of order k
}
