% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mu.R
\name{get_mu}
\alias{get_mu}
\title{Compute mu: the largest singular value of D multiplied by two times lambda}
\usage{
get_mu(n, degree, lambda)
}
\arguments{
\item{n}{length of the current observed daily counts}

\item{degree}{degree of the piecewise polynomial curve to be fitted,
e.g., degree = 0 corresponds to piecewise constant curves}

\item{lambda}{a parameter to balance the data fidelity and graphical
smoothness of fitted curves, e.g., a greater lambda results in better
smoothness}
}
\value{
mu
}
\description{
Compute mu: the largest singular value of D multiplied by two times lambda
}
\examples{
get_mu(5, 1, 0.1)
}
