% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.admm_rr}
\alias{plot.admm_rr}
\title{Plot a `admm_rr' class object}
\usage{
\method{plot}{admm_rr}(mod)
}
\arguments{
\item{mod}{a fitted admm_rr model}
}
\value{
a plot
}
\description{
Plot a `admm_rr' class object
}
