% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cv_poisson_rt}
\alias{plot.cv_poisson_rt}
\title{Plot cv_poisson_rt}
\usage{
\method{plot}{cv_poisson_rt}(x, plot_Rt = FALSE, which_lambda = NULL, ...)
}
\arguments{
\item{x}{result of cv_estimate_rt of class `cv_poisson_rt`}

\item{plot_Rt}{plot cross-validation scores only if set to `FALSE`; plot
Rt and specify which Rt to generate with the `which_lambda` parameter if set
to `TRUE`}

\item{which_lambda}{select which Rt's to plot, if `plot_Rt == TRUE`. If not
provided, plot R estimated with the optimal lambda. Any lambda provided must
match the values used in generating the Rt's.}

\item{...}{Not used.}
}
\value{
plot of cv scores
}
\description{
Plot cv_poisson_rt
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
cv <- cv_estimate_rt(y, degree = 3, fold = 2, nsol=30)
plot(cv)
plot(cv, plot_Rt = TRUE)
plot(cv, plot_Rt = TRUE, which_lambda = cv$lambda[1])
}
