% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cv_result}
\alias{plot.cv_result}
\title{Plot cv_result}
\usage{
\method{plot}{cv_result}(x, ...)
}
\arguments{
\item{x}{result of cv_estimate_rt of class `cv_result`}

\item{...}{}
}
\value{
plot of cv scores
}
\description{
Plot cv_result
}
\examples{
cv <- cv_estimate_rt(c(1:20), degree = 2, fold = 2, nsol=100)
plot(cv)
}
