% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.summary.admm_rr}
\alias{plot.summary.admm_rr}
\title{Plot summary of `admm_rr` models}
\usage{
\method{plot}{summary.admm_rr}(x, ...)
}
\arguments{
\item{x}{summary of `admm_rr` models}

\item{...}{.}
}
\value{
a figure
}
\description{
Plot summary of `admm_rr` models
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod = admm_solver(
  current_counts = y, weighted_past_counts = rep(1, 10), degree = 1,
  init = admm_initializer(current_counts = y,
  weighted_past_counts = rep(1, 10), degree = 1)
)
plot(summary(mod))
}
