% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.summary.cv_result}
\alias{plot.summary.cv_result}
\title{Plot of `summary.cv_result` object}
\usage{
\method{plot}{summary.cv_result}(cv_result, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot of `summary.cv_result` object
}
