% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.summary.poisson_rt}
\alias{plot.summary.poisson_rt}
\title{Plot predicted observed_count and estimated Rt from `summary(poisson_rt)` object}
\usage{
\method{plot}{summary.poisson_rt}(summary, ...)
}
\arguments{
\item{summary}{summary of `poisson_rt` models, of class `summary.poisson_rt`}

\item{...}{.}
}
\value{
Figure with two panels. Top panel shows the predicted observed_counts
calculated from \eqn{weighted_count * Rt_estim} against the true observed_counts.
Bottom panel shows the estimated Rt through observation time
}
\description{
Plot predicted observed_count and estimated Rt from `summary(poisson_rt)` object
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod <- estimate_rt(observed_counts = y, degree = 1, lambda = 0.001)
plot(summary(mod))
}
