% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.poisson_rt}
\alias{print.poisson_rt}
\title{Print `poisson_rt` object}
\usage{
\method{print}{poisson_rt}(x, \dots)
}
\arguments{
\item{x}{output of function `estimate_rt` of class `poisson_rt`}

\item{...}{further arguments passed to or from other methods.}
}
\value{
status of the `poisson_rt` object
}
\description{
Print `poisson_rt` object
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
out <- estimate_rt(y, lambda = log(c(1.1,1.3,1.5)))
out
}
