% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.poisson_rt}
\alias{summary.poisson_rt}
\title{Summarize results of a poisson_rt model}
\usage{
\method{summary}{poisson_rt}(object, ...)
}
\arguments{
\item{object}{output of `estimate_rt` function with class `poisson_rt`}

\item{...}{.}
}
\value{
a data table of estimates and a logical value of convergence.
The data table columns include the current observed daily counts (Signal),
the estimated reproduction rate (R_rate), and the estimated Poisson mean
parameter (pois_mean)
}
\description{
Summarize results of a poisson_rt model
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod = estiamte_rt(observed_counts = y, degree = 1, lambda = 0.001)
summary(mod)
}
