% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_solver.R
\name{admm_initializer}
\alias{admm_initializer}
\title{ADMM initialization}
\usage{
admm_initializer(y, x, k, z = NULL, u = NULL)
}
\arguments{
\item{y}{observed daily infection count}

\item{x}{scales of reproduction rate}

\item{k}{order of divided difference matrix D}

\item{z}{auxiliary variable}

\item{u}{dual variable}
}
\value{
a list of model initialization
}
\description{
ADMM initialization
}
