% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_solver.R
\name{admm_initializer}
\alias{admm_initializer}
\title{ADMM initialization}
\usage{
admm_initializer(
  current_counts,
  degree,
  weighted_past_counts = NULL,
  primal_var = NULL,
  auxi_var = NULL,
  dual_var = NULL,
  rho = -1,
  rho_adjust = -1,
  tolerance = 1e-04,
  verbose = 0
)
}
\arguments{
\item{current_counts}{the current daily infection counts}

\item{degree}{degree of the piecewise polynomial curves to be fitted,
e.g., degree = 0 corresponds to piecewise constant curves}

\item{weighted_past_counts}{the weighted sum of past infection counts with
corresponding serial interval functions (or its Gamma approximation) as
weights}

\item{primal_var}{the initial value of log(Rt)}

\item{auxi_var}{auxiliary variable in the ADMM algorithm}

\item{dual_var}{dual variable in the ADMM algorithm}
}
\value{
a list of model parameters with class `admm_initializer`
}
\description{
We should convert this into an S3 method so that we can pass in a
vector of counts or an admm_initializer object (and overwrite)
}
