% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_solver.R
\name{admm_solver}
\alias{admm_solver}
\title{ADMM solver}
\usage{
admm_solver(
  y,
  x,
  k,
  lambda = 0.01,
  mu = NULL,
  tol = 0.001,
  maxiter = 1000L,
  init = admm_initializer(y, x, k)
)
}
\arguments{
\item{y}{observed daily infection count}

\item{x}{scales of reproduction rate}

\item{k}{order of divided difference matrix}

\item{lambda}{Lagrangian multiplier}

\item{mu}{a hyper parameter depending on order, problem size, and lambda}

\item{tol}{tolerance of convergence of primal & dual residuals}

\item{maxiter}{maximal number of iteration}

\item{init}{ADMM initializer}
}
\value{
y: observed daily infection count

x: scales of reproduction rate

R_rate: reproduction rate

convr: if the model converges (convr==TRUE) or not (convr==FALSE)
}
\description{
ADMM solver
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
admm_solver(
  y = y, x = rep(1, 10), k = 1,
  init = admm_initializer(y = y, x = rep(1, 10), k = 1)
)
}
