% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_solver.R
\name{admm_solver}
\alias{admm_solver}
\title{Alternating Direction Method of Multiplier (ADMM) solver for smoothness penalized Poisson regression}
\usage{
admm_solver(
  current_counts,
  degree,
  dist_gamma = c(1, 1),
  x = double(0),
  lambda = double(0),
  nsol = 100L,
  lambdamin = -1,
  lambdamax = -1,
  lambda_min_ratio = 1e-04,
  maxiter = 10000,
  init = NULL
)
}
\arguments{
\item{current_counts}{the current daily infection counts}

\item{degree}{degree of the piecewise polynomial curve to be fitted,
e.g., degree = 0 corresponds to a piecewise constant curve}

\item{dist_gamma}{shape and scale parameter of the discretized Gamma distribution,
representing the serial interval distribution, which indicate how infectious someone is if they are infected a given days ago}

\item{x}{the observation position}

\item{lambda}{a parameter to balance the data fidelity and graphical
smoothness of fitted curves; a greater lambda results in a smoother curve}

\item{nsol}{number of lambdas to generate, if lambda is not specified}

\item{lambdamin}{If lambda is not specified, the program will generate a sequence of lambda, with lambdamin being the smallest lambda value}

\item{lambdamax}{If lambda is not specified, the program will generate a sequence of lambda, with lambdamin being the largest lambda value}

\item{lambda_min_ratio}{If lambda is not specified, and if lambdamin is not specified, the program will generate a lambdamin
by lambdamax * lambda_min_ratio}

\item{maxiter}{maximal number of iteration}

\item{init}{a list of model initialization of class `admm_initializer`}
}
\value{
current_counts the current daily infection counts

weighted_past_counts the weighted sum of past infection counts

R_rate: the estimated effective reproduction rate

convr: if the model converges `convr==TRUE` or not `convr==FALSE`
}
\description{
Alternating Direction Method of Multiplier (ADMM) to solve for a smoothness penalized Poisson regression.

The Effective Reproduction Number \eqn{R_t} of an infectious disease can be estimated by solving the smoothnesss penalized Poisson
regression of the form:

\eqn{R_t = argmin_{\theta} (\frac{1}{n} \sum_{i=1}^n e^{\theta_i} - y_i\theta_i) + \lambda||D^{(k)}\theta||_1}

where \eqn{y_i} is the observed case count at day \eqn{i}, \eqn{\theta_i = \sum_{a=1}y_{a}w_{t-a}} is the weighted past count at day \eqn{i},
\eqn{\lambda} is the smoothness penalty, and \eqn{D^{(k)}} is the \eqn{k}-th difference matrix
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
admm_solver(
  current_counts = y, weighted_past_counts = rep(1, 10), degree = 1,
  init = admm_initializer(current_counts = y,
  weighted_past_counts = rep(1, 10), degree = 1)
)
}
