% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rt.R
\name{configure_rt_admm}
\alias{configure_rt_admm}
\title{Rt estimation algorithm configuration}
\usage{
configure_rt_admm(
  observed_counts,
  degree,
  weighted_past_counts = NULL,
  primal_var = NULL,
  auxi_var = NULL,
  dual_var = NULL,
  rho = -1,
  rho_adjust = -1,
  tolerance = 1e-04,
  verbose = 0
)
}
\arguments{
\item{observed_counts}{vector of the observed daily infection counts}

\item{degree}{Integer. Degree of the piecewise polynomial curve to be
estimated. Ror example, `degree = 0` corresponds to a piecewise constant
curve.}

\item{weighted_past_counts}{the weighted sum of past infections counts with
corresponding serial interval functions (or its Gamma approximation) as
weights}

\item{primal_var}{initial values of log(Rt)}

\item{auxi_var}{auxiliary variable in the ADMM algorithm}

\item{dual_var}{dual variable in the ADMM algorithm}

\item{rho}{admm configuration parameter}

\item{rho_adjust}{admm configuration parameter}

\item{tolerance}{threshold to assess convergence}

\item{verbose}{control printing during fitting algorithm}
}
\value{
a list of model parameters with class `rt_admm_configuration`
}
\description{
We should convert this into an S3 method so that we can pass in a
vector of counts or an admm_initializer object (and overwrite)
}
