% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rt.R
\name{configure_rt_admm}
\alias{configure_rt_admm}
\title{Rt estimation algorithm configuration}
\usage{
configure_rt_admm(
  observed_counts,
  degree,
  weighted_past_counts = NULL,
  primal_var = NULL,
  auxi_var = NULL,
  dual_var = NULL,
  rho = -1,
  rho_adjust = -1,
  alpha = 0.5,
  gamma = 0.9,
  tolerance = 1e-04,
  verbose = 0
)
}
\arguments{
\item{observed_counts}{vector of the observed daily infection counts}

\item{degree}{Integer. Degree of the piecewise polynomial curve to be
estimated. For example, `degree = 1` corresponds to a piecewise constant
curve.}

\item{weighted_past_counts}{the weighted sum of past infections counts with
corresponding serial interval functions (or its Gamma approximation) as
weights}

\item{primal_var}{initial values of log(Rt)}

\item{auxi_var}{auxiliary variable in the ADMM algorithm}

\item{dual_var}{dual variable in the ADMM algorithm}

\item{rho}{double. An ADMM parameter; coefficient of augmented term in the
Lagrangian function.}

\item{rho_adjust}{double. An ADMM parameter; adjusted coefficient of
augmented term in the Lagrangian function.}

\item{alpha}{Double. A parameter adjusting upper bound in line search algorithm
in `irls_admm` algorithm.}

\item{gamma}{Double. A parameter adjusting step size in line search algorithm
in `irls_admm` algorithm.}

\item{tolerance}{double. Tolerance of ADMM convergence.}

\item{verbose}{integer.}
}
\value{
a list of model parameters with class `rt_admm_configuration`
}
\description{
We should convert this into an S3 method so that we can pass in a
vector of counts or an rt_admm_configuration object (and overwrite)
}
