% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rt.R
\name{configure_rt_admm}
\alias{configure_rt_admm}
\title{Rt estimation algorithm configuration}
\usage{
configure_rt_admm(
  observed_counts,
  korder,
  weighted_past_counts = NULL,
  rho = -1,
  alpha = 0.5,
  gamma = 0.9,
  tolerance = 1e-04,
  maxiter_newton = 50L,
  maxiter_line = 20L,
  verbose = 0
)
}
\arguments{
\item{observed_counts}{vector of the observed daily infection counts}

\item{korder}{Integer. Degree of the piecewise polynomial curve to be
estimated. For example, `korder = 0` corresponds to a piecewise constant
curve.}

\item{weighted_past_counts}{the weighted sum of past infections counts with
corresponding serial interval functions (or its Gamma approximation) as
weights}

\item{rho}{Double. An ADMM parameter; coefficient of augmented term in the
Lagrangian function.}

\item{alpha}{Double. A parameter adjusting upper bound in line search algorithm
in `prox_newton` algorithm.}

\item{gamma}{Double. A parameter adjusting step size in line search algorithm
in `prox_newton` algorithm.}

\item{tolerance}{Double. Tolerance of ADMM convergence.}

\item{maxiter_newton}{Integer. Maximum number of iterations for the outer
Newton iteration.}

\item{maxiter_line}{Integer. Maximum number of iterations for the linesearch
algorithm in the proximal Newton method.}

\item{verbose}{Integer.}
}
\value{
a list of model parameters with class `rt_admm_configuration`
}
\description{
Rt estimation algorithm configuration
}
