% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_estimate_rt.R
\name{cv_estimate_rt}
\alias{cv_estimate_rt}
\title{Leave-kth-out cross validation for choosing a optimal parameter lambda}
\usage{
cv_estimate_rt(
  observed_counts,
  degree = 3L,
  dist_gamma = c(2.5, 2.5),
  fold = 3,
  x = NULL,
  lambda = NULL,
  ...
)
}
\arguments{
\item{observed_counts}{vector of the observed daily infection counts}

\item{degree}{Integer. Degree of the piecewise polynomial curve to be
estimated. For example, `degree = 1` corresponds to a piecewise constant
curve.}

\item{dist_gamma}{Vector of length 2. These are the shape and scale for the
assumed serial interval distribution. Roughly, this distribution describes
the probability of an infectious individual infecting someone else after
some period of time after having become infectious.
As in most literature, we assume that this interval follows a gamma
distribution with some shape and scale.}

\item{fold}{Integer. This the number of folds to conduct the leave-kth-out
cross validation. For leave-kth-out cross validation, every kth
observed_counts and their corresponding position (evenly or unevenly
spaced) are placed into the same fold. The first and last observed_counts are
not assigned to any folds. Smallest allowable value if `fold = 2`. It is
generally not recommended to set `fold` to a large number}

\item{x}{a vector of positions at which the counts have been observed. In an
ideal case, we would observe data at regular intervals (e.g. daily or
weekly) but this may not always be the case.}

\item{lambda}{Vector. A user supplied sequence of tuning parameters which
determines the balance between data fidelity and
smoothness of the estimated Rt; larger `lambda` results in a smoother
estimate. The default, `NULL`
results in an automatic computation based on `nlambda`, the largest value
of `lambda` that would a maximally smooth estimate, and `lambda_min_ratio`.
Supplying a value of `lambda` overrides
this behaviour. It is likely better to supply a
decreasing sequence of `lambda` values than a single (small) value. If
supplied, the user-defined `lambda` sequence is automatically sorted in
decreasing order.}
}
\value{
An object with S3 class `"cv_result"`. Among the list components:
* `cv_scores` leave-kth-out cross validation scores
* `optimal_lambda` lambda which achieved the optimal cv_scores from  cross
validation
* `optimal_Rt` the estimated effective reproduction rate, estimated with the
optimal lambda
* `observed_counts` vector of the observed daily infection counts
* `weighted_past_counts` the weighted sum of past infections counts with
  corresponding serial interval functions (or its Gamma approximation) as
  weights
}
\description{
Leave-kth-out cross validation for choosing a optimal parameter lambda
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
cv <- cv_estimate_rt(y, degree = 3, fold = 2, nsol=30)
cv
}
