% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_gamma.R
\name{discretize_gamma}
\alias{discretize_gamma}
\title{discretized Gamma distribution given shape and scale parameters. This is intended to be the serial interval distribution of the infectious disease.}
\usage{
discretize_gamma(n, shape, scale)
}
\arguments{
\item{n, }{length of past observations}

\item{shape}{shape parameter of the Gamma distribution}

\item{scale}{scale parameter of the Gamma distribution}
}
\value{
probability mass of discretized gamma distribution in a vector of size n-1
}
\description{
discretized Gamma distribution given shape and scale parameters. This is intended to be the serial interval distribution of the infectious disease.
}
\examples{
discretize_gamma(30, shape = 2, scale = 2)
}
