% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_calculator.R
\name{fill_case_counts}
\alias{fill_case_counts}
\title{Fill case counts for uneven spacing}
\usage{
fill_case_counts(x, observed_counts)
}
\arguments{
\item{x}{a vector of positions at which the counts have been observed. In an
ideal case, we would observe data at regular intervals (e.g. daily or
weekly) but this may not always be the case.}

\item{observed_counts}{vector of the observed daily infection counts}
}
\description{
Fill case counts for uneven spacing
}
