% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fold_calculator}
\alias{fold_calculator}
\title{Helper function. Calculate the fold index for each `observed_counts`}
\usage{
fold_calculator(n, fold)
}
\arguments{
\item{n}{Integer. Total number of `observed_counts`}

\item{fold}{Integer. Number of folds for cross validation. In cross validation,
data are distributed into folds. Each fold during cross-validation is left as
the left-out data to be tested on after the model is fitted with the rest of
the set.}
}
\value{
a vector of fold index at which the counts are distributed into folds
}
\description{
Helper function. Calculate the fold index for each `observed_counts`
}
