% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildD.R
\name{get_mu}
\alias{get_mu}
\title{Compute mu}
\usage{
get_mu(n, k, lambda)
}
\arguments{
\item{n}{length of observed sequence}

\item{k}{order of divided difference}

\item{lambda}{tuning parameter}
}
\value{
mu
}
\description{
Compute mu
}
\examples{
get_mu(5, 1, 0.1)
}
