% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_rt.R
\name{interpolate_rt}
\alias{interpolate_rt}
\title{Interpolate (or extrapolate) Rt estimates to intermediate design points}
\usage{
interpolate_rt(object, xout, ...)
}
\arguments{
\item{object}{A fitted object produced by \code{estimate_rt()} or \code{cv_estimate_rt()}.}

\item{xout}{a vector of new positions at which Rt should be produced,
but where counts may not have been observed.}

\item{...}{additional arguments passed to methods.}
}
\value{
A vector or matrix of interpolated Rt estimates.
}
\description{
Interpolate (or extrapolate) Rt estimates to intermediate design points
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15) * 500 + 1))
out <- estimate_rt(y)

# originally estimated at
out$x

# get the Rt at 3 new points (for all estimated lambdas)
int <- interpolate_rt(out, c(10.5, 11.5, 12.5))

# get the Rt at a single value of lambda
interpolate_rt(out, c(10.5, 11.5, 12.5), lambda = out$lambda[20])

}
