% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_rt.R, R/methods.R
\name{interpolate_rt}
\alias{interpolate_rt}
\alias{interpolate_rt.poisson_rt}
\title{Interpolate (or extrapolate) Rt estimates to intermediate design points}
\usage{
interpolate_rt(object, xout, ...)

\method{interpolate_rt}{poisson_rt}(object, xout, lambda = NULL, ...)
}
\arguments{
\item{object}{A fitted object produced by \code{estimate_rt()} or \code{cv_estimate_rt()}.}

\item{xout}{a vector of new positions at which Rt should be produced,
but where counts may not have been observed.}

\item{...}{additional arguments passed to methods.}
}
\value{
A vector or matrix of interpolated Rt estimates.
}
\description{
Interpolate (or extrapolate) Rt estimates to intermediate design points
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15) * 500 + 1))
out <- estimate_rt(y, nsol = 10)
interpolate_rt(out, xout = c(1.5, 2.5))
}
