% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{match_lambda}
\alias{match_lambda}
\title{Find index of elements from a smaller list in a bigger list}
\usage{
match_lambda(which_lambda, lambda)
}
\arguments{
\item{which_lambda}{list of elements to be found indices from `lambda`}

\item{lambda}{list where elements are to be found}
}
\value{
list of index of the elements in `which_lambda` from `lambda`
}
\description{
Find index of elements from a smaller list in a bigger list
}
\examples{
lambda <- c(1:20)
which_lambda <- c(1,3,6)
which_lambda_notin(lambda, which_lambda)
}
