% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{match_lambda}
\alias{match_lambda}
\title{Find indices of a list of elements from another list}
\usage{
match_lambda(which_lambda, lambda)
}
\arguments{
\item{which_lambda}{list of elements whose indices are to be found from `lambda`}

\item{lambda}{list where the indices of elements are to be found from}
}
\value{
list of index of the elements in `which_lambda` from `lambda`
}
\description{
Find indices of every elements from `which_lambda` in `lambda`. Return
the indices if all elements in `which_lambda` are in `lambda`. Return
indices and warning if some elements in `which_lambda` are not in `lambda`.
Abort the program if no element in `which_lambda` is in `lambda`
}
\examples{
lambda <- c(1:20)
which_lambda <- c(1,3,6)
match_lambda(lambda, which_lambda)
}
