% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.admm_rr}
\alias{plot.admm_rr}
\title{Plot a `admm_rr' class object}
\usage{
\method{plot}{admm_rr}(x, y, ...)
}
\arguments{
\item{x}{time}

\item{y}{signal; estimated poisson mean}

\item{...}{.}
}
\value{
a plot
}
\description{
Plot a `admm_rr' class object
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod = admm_solver(
  y = y, x = rep(1, 10), k = 1,
  init = admm_initializer(y = y, x = rep(1, 10), k = 1)
)
res <- summary(mod)$Results
plot(res$pois_para, y)
}
