% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.poisson_rt}
\alias{plot.poisson_rt}
\title{Plot `poisson_rt` object}
\usage{
\method{plot}{poisson_rt}(x, which_lambda = NULL, ...)
}
\arguments{
\item{x}{output of function `estimate_rt` of class `poisson_rt`}

\item{which_lambda}{select which Rt's to plot. If no lambdas are provided,
all Rt's are plotted. Lambdas provided must match the lambda used in
generating the Rt's.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
plot of all or selected Rt from an object of class `poisson_rt`
}
\description{
Plot `poisson_rt` object
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
out <- estimate_rt(y, lambda = log(c(1.1,1.3,1.5)))
plot(out)
}
