% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.poisson_rt}
\alias{plot.poisson_rt}
\title{Plot `poisson_rt` object}
\usage{
\method{plot}{poisson_rt}(x, ...)
}
\arguments{
\item{x}{output of function `estimate_rt` of class `poisson_rt`}

\item{...}{}
}
\description{
Plot `poisson_rt` object
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod <- estimate_rt(observed_counts = y, degree = 1, lambda = 0.001)
plot(mod)
}
