% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.summary.poisson_rt}
\alias{plot.summary.poisson_rt}
\title{Plot summary of `poisson_rt` models}
\usage{
\method{plot}{summary.poisson_rt}(x, ...)
}
\arguments{
\item{x}{summary of `poisson_rt` models}

\item{...}{.}
}
\value{
a figure
}
\description{
Plot summary of `poisson_rt` models
}
\examples{
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod <- estimate_rt(
  observed_counts = y, dist_gamma=c(1,1), degree = 1, lambda = .1,
  algo = "linear_admm",
  init = rt_admm_configuration(y, degree = 1)
)
plot(summary(mod))
}
