% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.summary.poisson_rt}
\alias{plot.summary.poisson_rt}
\title{Plot summary of `poisson_rt` models}
\usage{
\method{plot}{summary.poisson_rt}(summary, ...)
}
\arguments{
\item{...}{.}

\item{x}{summary of `poisson_rt` models}
}
\value{
a figure
}
\description{
Plot summary of `poisson_rt` models
}
\examples{
TODO: change this example
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod = admm_solver(
  current_counts = y, weighted_past_counts = rep(1, 10), degree = 1,
  init = admm_initializer(current_counts = y,
  weighted_past_counts = rep(1, 10), degree = 1)
)
plot(summary(mod))
}
