% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{pred_kth_rt}
\alias{pred_kth_rt}
\title{Helper function. Calculate Rt at hold-out fold}
\usage{
pred_kth_rt(rt, n, train_idx, test_idx, train_x, test_x)
}
\arguments{
\item{rt}{an nlambda by (n-n_holdout) matrix. Rt estimated based on all lambdas for a
particular fold.}

\item{test_x}{}
}
\description{
Helper function. Calculate Rt at hold-out fold
}
