% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_estimate_rt.R
\name{pred_kth_rt}
\alias{pred_kth_rt}
\title{Helper function. Calculate Rt at hold-out set}
\usage{
pred_kth_rt(rt, n, train_idx, test_idx, train_x, test_x)
}
\arguments{
\item{rt}{Matrix. Rt estimation at observed time points in training set
for all values of lambdas}

\item{n}{Integer. Number of total observations}

\item{train_idx}{vector of Integers. Index of \code{observed_counts} and \code{x} to be
assigned to the training set}

\item{test_idx}{vector of Integers. Index of \code{observed_counts} and \code{x} to be
assigned to the testing set}

\item{train_x}{vector of Integers. Location of the \code{observed_counts} in the
training set}

\item{test_x}{vector of Integers. Location of the \code{observed_counts} in the
testing set}
}
\value{
Predicted Rt at the hold-out set
}
\description{
Helper function. Calculate Rt at hold-out set
}
