% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.admm_rr}
\alias{summary.admm_rr}
\title{Summarize results of a admm_rr model}
\usage{
\method{summary}{admm_rr}(mod)
}
\arguments{
\item{mod}{a fitted admm_rr model}
}
\value{
a data table of estimates and a logical value of convergence
}
\description{
Summarize results of a admm_rr model
}
