% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cv_result}
\alias{summary.cv_result}
\title{Summary of `cv_result` object}
\usage{
\method{summary}{cv_result}(object, ...)
}
\arguments{
\item{...}{}
}
\description{
Summary of `cv_result` object
}
