% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.poisson_rt}
\alias{summary.poisson_rt}
\title{Summary of the `poisson_rt` object}
\usage{
summary.poisson_rt(object, ...)
}
\arguments{
\item{...}{}
}
\description{
Summary of the `poisson_rt` object
}
