% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.poisson_rt}
\alias{summary.poisson_rt}
\title{Summarize results of a poisson_rt model}
\usage{
\method{summary}{poisson_rt}(object, ...)
}
\arguments{
\item{object}{a fitted model of class `poisson_rt`}

\item{...}{.}
}
\value{
a data table of estimates and a logical value of convergence.
The data table columns include the current observed daily counts (Signal),
the estimated reproduction rate (R_rate), and the estimated Poisson mean
parameter (pois_mean)
}
\description{
Summarize results of a poisson_rt model
}
\examples{
TODO: Need to change example
y <- c(rev(seq(2, 6, by = 1)), seq(2, 6, by = 1))
mod = admm_solver(
  current_counts = y, weighted_past_counts = rep(1, 10), degree = 1,
  init = admm_initializer(current_counts = y,
  weighted_past_counts = rep(1, 10), degree = 1)
)
summary(mod)
}
