% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.poisson_rt}
\alias{summary.poisson_rt}
\title{Summary of the `poisson_rt` object}
\usage{
summary.poisson_rt(object, ...)
}
\arguments{
\item{object}{output of function `estimate_rt` of class `poisson_rt`}

\item{...}{}
}
\value{
summary of the `poisson_rt` object in a table
}
\description{
summarize the `poisson_rt` object in a table
}
\examples{
y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
out <- estimate_rt(y, lambda = log(c(1.1,1.3,1.5)))
summary(out)
}
